feat: move binding rituals category into a category provider
