+++ src/main/java/com/klikli_dev/occultism/datagen/OccultismBookProvider.java
@@ -14,6 +14,7 @@ import com.klikli_dev.modonomicon.api.datagen.book.condition.BookModLoadedCondit
 import com.klikli_dev.modonomicon.api.datagen.book.condition.BookTrueConditionModel;
 import com.klikli_dev.modonomicon.api.datagen.book.page.*;
 import com.klikli_dev.occultism.Occultism;
+import com.klikli_dev.occultism.datagen.book.BindingRitualsCategory;
 import com.klikli_dev.occultism.datagen.book.FamiliarRitualsCategory;
 import com.klikli_dev.occultism.integration.modonomicon.pages.*;
 import com.klikli_dev.occultism.registry.OccultismBlocks;
@@ -50,7 +51,7 @@ public class OccultismBookProvider extends SingleBookSubProvider {
 
         var summoningRitualsCategory = this.add(this.makeSummoningRitualsSubcategory().withSortNumber(sortNum++));
         var possessionRitualsCategory = this.add(this.makePossessionRitualsSubcategory().withSortNumber(sortNum++));
-        var craftingRitualsCategory = this.add(this.makeCraftingRitualsSubcategory().withSortNumber(sortNum++));
+        var craftingRitualsCategory = this.add(new BindingRitualsCategory(this).generate().withSortNumber(sortNum++));
         var familiarRitualsCategory = this.add(new FamiliarRitualsCategory(this).generate().withSortNumber(sortNum++));
 
         var pentaclesCategory = this.add(this.makePentaclesCategory().withSortNumber(sortNum++));
@@ -88,6 +89,16 @@ public class OccultismBookProvider extends SingleBookSubProvider {
                 .withAutoAddReadConditions(true)
                 .withAllowOpenBooksWithInvalidLinks(true)
                 ;
+    }
+
+    private BookEntryModel makeReturnToRitualsEntry(CategoryEntryMap entryMap, char icon) {
+        this.context().entry("return_to_rituals");
+
+        return BookEntryModel.create(this.modLoc(this.context().categoryId() + "/" + this.context().entryId()), this.context().entryName())
+                .withIcon(this.modLoc("textures/gui/book/robe.png"))
+                .withCategoryToOpen(this.modLoc("rituals"))
+                .withEntryBackground(1, 2)
+                .withLocation(entryMap.get(icon));
     }
 
     //region Getting Started
@@ -3241,15 +3252,7 @@ public class OccultismBookProvider extends SingleBookSubProvider {
                 .withLocation(entryMap.get(icon));
     }
 
-    private BookEntryModel makeReturnToRitualsEntry(CategoryEntryMap entryMap, char icon) {
-        this.context().entry("return_to_rituals");
 
-        return BookEntryModel.create(this.modLoc(this.context().categoryId() + "/" + this.context().entryId()), this.context().entryName())
-                .withIcon(this.modLoc("textures/gui/book/robe.png"))
-                .withCategoryToOpen(this.modLoc("rituals"))
-                .withEntryBackground(1, 2)
-                .withLocation(entryMap.get(icon));
-    }
     //endregion
 
     //region Summoning Rituals
@@ -4046,629 +4049,6 @@ public class OccultismBookProvider extends SingleBookSubProvider {
     }
     //endregion
 
-    //region Crafting Rituals
-    private BookCategoryModel makeCraftingRitualsSubcategory() {
-        this.context().category("crafting_rituals");
-
-        var entryMap = ModonomiconAPI.get().getEntryMap();
-        entryMap.setMap(
-                "___________________________",
-                "_______b_e_x_p_q___________",
-                "___________________________",
-                "_______d_h_c__w____________",
-                "___________________________",
-                "___9_0_____________________",
-                "___________________________",
-                "_______f_z_a__g____________",
-                "___________________________",
-                "___________n_m_o___________",
-                "___________________________",
-                "___________i_j_k_l_________",
-                "___________________________"
-        );
-
-        var overview = this.makeCraftingRitualsOverviewEntry(entryMap, '0');
-        var returnToRituals = this.makeReturnToRitualsEntry(entryMap, '9');
-        returnToRituals.withParent(BookEntryParentModel.create(overview.getId()));
-        returnToRituals.withCondition(BookTrueConditionModel.create());
-
-        var craftInfusedPickaxe = this.makeCraftInfusedPickaxeEntry(entryMap, 'd');
-        craftInfusedPickaxe.withParent(BookEntryParentModel.create(overview.getId()));
-        var craftDimensionalMineshaft = this.makeCraftDimensionalMineshaftEntry(entryMap, 'b');
-        craftDimensionalMineshaft.withParent(BookEntryParentModel.create(craftInfusedPickaxe.getId()));
-        var craftFoliotMiner = this.makeCraftFoliotMinerEntry(entryMap, 'e');
-        craftFoliotMiner.withParent(BookEntryParentModel.create(craftDimensionalMineshaft.getId()));
-        var craftDjinniMiner = this.makeCraftDjinniMinerEntry(entryMap, 'x');
-        craftDjinniMiner.withParent(BookEntryParentModel.create(craftFoliotMiner.getId()));
-        var craftAfritMiner = this.makeCraftAfritMinerEntry(entryMap, 'p');
-        craftAfritMiner.withParent(BookEntryParentModel.create(craftDjinniMiner.getId()));
-        var craftMaridMiner = this.makeCraftMaridMinerEntry(entryMap, 'q');
-        craftMaridMiner.withParent(BookEntryParentModel.create(craftAfritMiner.getId()));
-
-        var craftStorageSystem = this.makeCraftStorageSystemEntry(entryMap, 'z');
-        craftStorageSystem.withParent(BookEntryParentModel.create(overview.getId()));
-
-        var craftDimensionalMatrix = this.makeCraftDimensionalMatrixEntry(entryMap, 'a');
-        craftDimensionalMatrix.withParent(BookEntryParentModel.create(craftStorageSystem.getId()));
-        var craftStorageControllerBase = this.makeCraftStorageControllerBaseEntry(entryMap, 'n');
-        craftStorageControllerBase.withParent(BookEntryParentModel.create(craftDimensionalMatrix.getId()));
-        var craftStabilizerTier1 = this.makeCraftStabilizerTier1Entry(entryMap, 'i');
-        craftStabilizerTier1.withParent(BookEntryParentModel.create(craftStorageControllerBase.getId()));
-        var craftStabilizerTier2 = this.makeCraftStabilizerTier2Entry(entryMap, 'j');
-        craftStabilizerTier2.withParent(BookEntryParentModel.create(craftStabilizerTier1.getId()));
-        var craftStabilizerTier3 = this.makeCraftStabilizerTier3Entry(entryMap, 'k');
-        craftStabilizerTier3.withParent(BookEntryParentModel.create(craftStabilizerTier2.getId()));
-        var craftStabilizerTier4 = this.makeCraftStabilizerTier4Entry(entryMap, 'l');
-        craftStabilizerTier4.withParent(BookEntryParentModel.create(craftStabilizerTier3.getId()));
-
-        var craftStableWormhole = this.makeCraftStableWormholeEntry(entryMap, 'm');
-        craftStableWormhole.withParent(BookEntryParentModel.create(craftStorageControllerBase.getId()));
-        var craftStorageRemote = this.makeCraftStorageRemoteEntry(entryMap, 'o');
-        craftStorageRemote.withParent(BookEntryParentModel.create(craftStableWormhole.getId()));
-
-        var craftOtherworldGoggles = this.makeCraftOtherworldGogglesEntry(entryMap, 'f');
-        craftOtherworldGoggles.withParent(BookEntryParentModel.create(overview.getId()));
-
-        var craftSatchel = this.makeCraftSatchelEntry(entryMap, 'g');
-        craftSatchel.withParent(BookEntryParentModel.create(overview.getId()));
-
-        var craftSoulGem = this.makeCraftSoulGemEntry(entryMap, 'h');
-        craftSoulGem.withParent(BookEntryParentModel.create(overview.getId()));
-        var craftFamiliarRing = this.makeCraftFamiliarRingEntry(entryMap, 'c');
-        craftFamiliarRing.withParent(BookEntryParentModel.create(craftSoulGem.getId()));
-
-        var craftWildTrim = this.makeCraftWildTrimEntry(entryMap, 'w');
-        craftWildTrim.withParent(BookEntryParentModel.create(overview.getId()));
-
-        //add true condition to all entries to enable them by default
-        overview.withCondition(BookTrueConditionModel.create());
-        craftInfusedPickaxe.withCondition(BookTrueConditionModel.create());
-        craftDimensionalMineshaft.withCondition(BookTrueConditionModel.create());
-        craftFoliotMiner.withCondition(BookTrueConditionModel.create());
-        craftDjinniMiner.withCondition(BookTrueConditionModel.create());
-        craftAfritMiner.withCondition(BookTrueConditionModel.create());
-        craftMaridMiner.withCondition(BookTrueConditionModel.create());
-        craftStorageSystem.withCondition(BookTrueConditionModel.create());
-        craftDimensionalMatrix.withCondition(BookTrueConditionModel.create());
-        craftStorageControllerBase.withCondition(BookTrueConditionModel.create());
-        craftStabilizerTier1.withCondition(BookTrueConditionModel.create());
-        craftStabilizerTier2.withCondition(BookTrueConditionModel.create());
-        craftStabilizerTier3.withCondition(BookTrueConditionModel.create());
-        craftStabilizerTier4.withCondition(BookTrueConditionModel.create());
-        craftStableWormhole.withCondition(BookTrueConditionModel.create());
-        craftStorageRemote.withCondition(BookTrueConditionModel.create());
-        craftOtherworldGoggles.withCondition(BookTrueConditionModel.create());
-        craftSatchel.withCondition(BookTrueConditionModel.create());
-        craftSoulGem.withCondition(BookTrueConditionModel.create());
-        craftFamiliarRing.withCondition(BookTrueConditionModel.create());
-        craftWildTrim.withCondition(BookTrueConditionModel.create());
-
-        return BookCategoryModel.create(this.modLoc(this.context().categoryId()), this.context().categoryName())
-                .withIcon(this.modLoc("textures/gui/book/infusion.png"))
-                .withShowCategoryButton(true)
-                .withEntries(
-                        overview,
-                        returnToRituals,
-                        craftStorageSystem,
-                        craftDimensionalMatrix,
-                        craftDimensionalMineshaft,
-                        craftInfusedPickaxe,
-                        craftFoliotMiner,
-                        craftDjinniMiner,
-                        craftAfritMiner,
-                        craftMaridMiner,
-                        craftOtherworldGoggles,
-                        craftSatchel,
-                        craftSoulGem,
-                        craftFamiliarRing,
-                        craftStabilizerTier1,
-                        craftStabilizerTier2,
-                        craftStabilizerTier3,
-                        craftStabilizerTier4,
-                        craftStableWormhole,
-                        craftStorageControllerBase,
-                        craftStorageRemote,
-                        craftWildTrim
-                );
-    }
-
-    private BookEntryModel makeCraftingRitualsOverviewEntry(CategoryEntryMap entryMap, char icon) {
-        this.context().entry("overview");
-
-        this.context().page("intro");
-        var intro = BookTextPageModel.create()
-                .withTitle(this.context().pageTitle())
-                .withText(this.context().pageText());
-
-        return BookEntryModel.create(this.modLoc(this.context().categoryId() + "/" + this.context().entryId()), this.context().entryName())
-                .withIcon(this.modLoc("textures/gui/book/infusion.png"))
-                .withLocation(entryMap.get(icon))
-                .withEntryBackground(0, 1)
-                .withPages(
-                        intro
-                );
-    }
-
-    private BookEntryModel makeCraftStorageSystemEntry(CategoryEntryMap entryMap, char icon) {
-        this.context().entry("craft_storage_system");
-
-        this.context().page("spotlight");
-        var spotlight = BookSpotlightPageModel.create()
-                .withItem(Ingredient.of(OccultismBlocks.STORAGE_CONTROLLER.get()))
-                .withText(this.context().pageText());
-
-        return BookEntryModel.create(this.modLoc(this.context().categoryId() + "/" + this.context().entryId()), this.context().entryName())
-                .withIcon(Items.CHEST)
-                .withLocation(entryMap.get(icon))
-                .withPages(
-                        spotlight
-                );
-    }
-
-
-    private BookEntryModel makeCraftDimensionalMatrixEntry(CategoryEntryMap entryMap, char icon) {
-        this.context().entry("craft_dimensional_matrix");
-
-        this.context().page("spotlight");
-        var spotlight = BookSpotlightPageModel.create()
-                .withItem(Ingredient.of(OccultismItems.DIMENSIONAL_MATRIX.get()))
-                .withText(this.context().pageText());
-
-        this.context().page("ritual");
-        var ritual = BookRitualRecipePageModel.create()
-                .withRecipeId1(this.modLoc("ritual/craft_dimensional_matrix"));
-
-        return BookEntryModel.create(this.modLoc(this.context().categoryId() + "/" + this.context().entryId()), this.context().entryName())
-                .withIcon(OccultismItems.DIMENSIONAL_MATRIX.get())
-                .withLocation(entryMap.get(icon))
-                .withPages(
-                        spotlight,
-                        ritual
-                );
-    }
-
-    private BookEntryModel makeCraftDimensionalMineshaftEntry(CategoryEntryMap entryMap, char icon) {
-        this.context().entry("craft_dimensional_mineshaft");
-
-        this.context().page("spotlight");
-        var spotlight = BookSpotlightPageModel.create()
-                .withItem(Ingredient.of(OccultismBlocks.DIMENSIONAL_MINESHAFT.get()))
-                .withText(this.context().pageText());
-
-        this.context().page("ritual");
-        var ritual = BookRitualRecipePageModel.create()
-                .withRecipeId1(this.modLoc("ritual/craft_dimensional_mineshaft"));
-
-        this.context().page("description");
-        var description = BookTextPageModel.create()
-                .withTitle(this.context().pageTitle())
-                .withText(this.context().pageText());
-
-        return BookEntryModel.create(this.modLoc(this.context().categoryId() + "/" + this.context().entryId()), this.context().entryName())
-                .withIcon(OccultismBlocks.DIMENSIONAL_MINESHAFT.get())
-                .withLocation(entryMap.get(icon))
-                .withPages(
-                        spotlight,
-                        ritual,
-                        description
-                );
-    }
-
-    private BookEntryModel makeCraftInfusedPickaxeEntry(CategoryEntryMap entryMap, char icon) {
-        this.context().entry("craft_infused_pickaxe");
-
-        this.context().page("spotlight");
-        var spotlight = BookSpotlightPageModel.create()
-                .withItem(Ingredient.of(OccultismItems.INFUSED_PICKAXE.get()))
-                .withText(this.context().pageText());
-
-        this.context().page("ritual");
-        var ritual = BookRitualRecipePageModel.create()
-                .withRecipeId1(this.modLoc("ritual/craft_infused_pickaxe"));
-
-        return BookEntryModel.create(this.modLoc(this.context().categoryId() + "/" + this.context().entryId()), this.context().entryName())
-                .withIcon(OccultismItems.INFUSED_PICKAXE.get())
-                .withLocation(entryMap.get(icon))
-                .withPages(
-                        spotlight,
-                        ritual
-                );
-    }
-
-    private BookEntryModel makeCraftStorageControllerBaseEntry(CategoryEntryMap entryMap, char icon) {
-        this.context().entry("craft_storage_controller_base");
-
-        this.context().page("spotlight");
-        var spotlight = BookSpotlightPageModel.create()
-                .withItem(Ingredient.of(OccultismBlocks.STORAGE_CONTROLLER_BASE.get()))
-                .withText(this.context().pageText());
-
-        this.context().page("ritual");
-        var ritual = BookRitualRecipePageModel.create()
-                .withRecipeId1(this.modLoc("ritual/craft_storage_controller_base"));
-
-        return BookEntryModel.create(this.modLoc(this.context().categoryId() + "/" + this.context().entryId()), this.context().entryName())
-                .withIcon(OccultismBlocks.STORAGE_CONTROLLER_BASE.get())
-                .withLocation(entryMap.get(icon))
-                .withPages(
-                        spotlight,
-                        ritual
-                );
-    }
-
-    private BookEntryModel makeCraftStabilizerTier1Entry(CategoryEntryMap entryMap, char icon) {
-        this.context().entry("craft_stabilizer_tier1");
-
-        this.context().page("spotlight");
-        var spotlight = BookSpotlightPageModel.create()
-                .withItem(Ingredient.of(OccultismBlocks.STORAGE_STABILIZER_TIER1.get()))
-                .withText(this.context().pageText());
-
-        this.context().page("ritual");
-        var ritual = BookRitualRecipePageModel.create()
-                .withRecipeId1(this.modLoc("ritual/craft_stabilizer_tier1"));
-
-        return BookEntryModel.create(this.modLoc(this.context().categoryId() + "/" + this.context().entryId()), this.context().entryName())
-                .withIcon(OccultismBlocks.STORAGE_STABILIZER_TIER1.get())
-                .withLocation(entryMap.get(icon))
-                .withPages(
-                        spotlight,
-                        ritual
-                );
-    }
-
-    private BookEntryModel makeCraftStabilizerTier2Entry(CategoryEntryMap entryMap, char icon) {
-        this.context().entry("craft_stabilizer_tier2");
-
-        this.context().page("spotlight");
-        var spotlight = BookSpotlightPageModel.create()
-                .withItem(Ingredient.of(OccultismBlocks.STORAGE_STABILIZER_TIER2.get()))
-                .withText(this.context().pageText());
-
-        this.context().page("ritual");
-        var ritual = BookRitualRecipePageModel.create()
-                .withRecipeId1(this.modLoc("ritual/craft_stabilizer_tier2"));
-
-        return BookEntryModel.create(this.modLoc(this.context().categoryId() + "/" + this.context().entryId()), this.context().entryName())
-                .withIcon(OccultismBlocks.STORAGE_STABILIZER_TIER2.get())
-                .withLocation(entryMap.get(icon))
-                .withPages(
-                        spotlight,
-                        ritual
-                );
-    }
-
-    private BookEntryModel makeCraftStabilizerTier3Entry(CategoryEntryMap entryMap, char icon) {
-        this.context().entry("craft_stabilizer_tier3");
-
-        this.context().page("spotlight");
-        var spotlight = BookSpotlightPageModel.create()
-                .withItem(Ingredient.of(OccultismBlocks.STORAGE_STABILIZER_TIER3.get()))
-                .withText(this.context().pageText());
-
-        this.context().page("ritual");
-        var ritual = BookRitualRecipePageModel.create()
-                .withRecipeId1(this.modLoc("ritual/craft_stabilizer_tier3"));
-
-        return BookEntryModel.create(this.modLoc(this.context().categoryId() + "/" + this.context().entryId()), this.context().entryName())
-                .withIcon(OccultismBlocks.STORAGE_STABILIZER_TIER3.get())
-                .withLocation(entryMap.get(icon))
-                .withPages(
-                        spotlight,
-                        ritual
-                );
-    }
-
-    private BookEntryModel makeCraftStabilizerTier4Entry(CategoryEntryMap entryMap, char icon) {
-        this.context().entry("craft_stabilizer_tier4");
-
-        this.context().page("spotlight");
-        var spotlight = BookSpotlightPageModel.create()
-                .withItem(Ingredient.of(OccultismBlocks.STORAGE_STABILIZER_TIER4.get()))
-                .withText(this.context().pageText());
-
-        this.context().page("ritual");
-        var ritual = BookRitualRecipePageModel.create()
-                .withRecipeId1(this.modLoc("ritual/craft_stabilizer_tier4"));
-
-        return BookEntryModel.create(this.modLoc(this.context().categoryId() + "/" + this.context().entryId()), this.context().entryName())
-                .withIcon(OccultismBlocks.STORAGE_STABILIZER_TIER4.get())
-                .withLocation(entryMap.get(icon))
-                .withPages(
-                        spotlight,
-                        ritual
-                );
-    }
-
-    private BookEntryModel makeCraftStableWormholeEntry(CategoryEntryMap entryMap, char icon) {
-        this.context().entry("craft_stable_wormhole");
-
-        this.context().page("spotlight");
-        var spotlight = BookSpotlightPageModel.create()
-                .withItem(Ingredient.of(OccultismBlocks.STABLE_WORMHOLE.get()))
-                .withText(this.context().pageText());
-
-        this.context().page("ritual");
-        var ritual = BookRitualRecipePageModel.create()
-                .withRecipeId1(this.modLoc("ritual/craft_stable_wormhole"));
-
-        return BookEntryModel.create(this.modLoc(this.context().categoryId() + "/" + this.context().entryId()), this.context().entryName())
-                .withIcon(OccultismBlocks.STABLE_WORMHOLE.get())
-                .withLocation(entryMap.get(icon))
-                .withPages(
-                        spotlight,
-                        ritual
-                );
-    }
-
-    private BookEntryModel makeCraftStorageRemoteEntry(CategoryEntryMap entryMap, char icon) {
-        this.context().entry("craft_storage_remote");
-
-        this.context().page("spotlight");
-        var spotlight = BookSpotlightPageModel.create()
-                .withItem(Ingredient.of(OccultismItems.STORAGE_REMOTE.get()))
-                .withText(this.context().pageText());
-
-        this.context().page("ritual");
-        var ritual = BookRitualRecipePageModel.create()
-                .withRecipeId1(this.modLoc("ritual/craft_storage_remote"));
-
-        return BookEntryModel.create(this.modLoc(this.context().categoryId() + "/" + this.context().entryId()), this.context().entryName())
-                .withIcon(OccultismItems.STORAGE_REMOTE.get())
-                .withLocation(entryMap.get(icon))
-                .withPages(
-                        spotlight,
-                        ritual
-                );
-    }
-
-    private BookEntryModel makeCraftFoliotMinerEntry(CategoryEntryMap entryMap, char icon) {
-        this.context().entry("craft_foliot_miner");
-
-        this.context().page("intro");
-        var intro = BookTextPageModel.create()
-                .withTitle(this.context().pageTitle())
-                .withText(this.context().pageText());
-
-        this.context().page("magic_lamp");
-        var lamp = BookTextPageModel.create()
-                .withTitle(this.context().pageTitle())
-                .withText(this.context().pageText());
-
-        this.context().page("magic_lamp_recipe");
-        var lampRecipe = BookCraftingRecipePageModel.create()
-                .withRecipeId1(this.modLoc("crafting/magic_lamp_empty"));
-
-        this.context().page("spotlight");
-        var spotlight = BookSpotlightPageModel.create()
-                .withItem(Ingredient.of(OccultismItems.MINER_FOLIOT_UNSPECIALIZED.get()))
-                .withText(this.context().pageText());
-
-        this.context().page("ritual");
-        var ritual = BookRitualRecipePageModel.create()
-                .withRecipeId1(this.modLoc("ritual/craft_miner_foliot_unspecialized"));
-
-        return BookEntryModel.create(this.modLoc(this.context().categoryId() + "/" + this.context().entryId()), this.context().entryName())
-                .withIcon(OccultismItems.MINER_FOLIOT_UNSPECIALIZED.get())
-                .withLocation(entryMap.get(icon))
-                .withPages(
-                        intro,
-                        lamp,
-                        lampRecipe,
-                        spotlight,
-                        ritual
-                );
-    }
-
-    private BookEntryModel makeCraftDjinniMinerEntry(CategoryEntryMap entryMap, char icon) {
-        this.context().entry("craft_djinni_miner");
-
-        this.context().page("spotlight");
-        var spotlight = BookSpotlightPageModel.create()
-                .withItem(Ingredient.of(OccultismItems.MINER_DJINNI_ORES.get()))
-                .withText(this.context().pageText());
-
-        this.context().page("ritual");
-        var ritual = BookRitualRecipePageModel.create()
-                .withRecipeId1(this.modLoc("ritual/craft_miner_djinni_ores"));
-
-        return BookEntryModel.create(this.modLoc(this.context().categoryId() + "/" + this.context().entryId()), this.context().entryName())
-                .withIcon(OccultismItems.MINER_DJINNI_ORES.get())
-                .withLocation(entryMap.get(icon))
-                .withPages(
-                        spotlight,
-                        ritual
-                );
-    }
-
-    private BookEntryModel makeCraftAfritMinerEntry(CategoryEntryMap entryMap, char icon) {
-        this.context().entry("craft_afrit_miner");
-
-        this.context().page("spotlight");
-        var spotlight = BookSpotlightPageModel.create()
-                .withItem(Ingredient.of(OccultismItems.MINER_AFRIT_DEEPS.get()))
-                .withText(this.context().pageText());
-
-        this.context().page("ritual");
-        var ritual = BookRitualRecipePageModel.create()
-                .withRecipeId1(this.modLoc("ritual/craft_miner_afrit_deeps"));
-
-        return BookEntryModel.create(this.modLoc(this.context().categoryId() + "/" + this.context().entryId()), this.context().entryName())
-                .withIcon(OccultismItems.MINER_AFRIT_DEEPS.get())
-                .withLocation(entryMap.get(icon))
-                .withPages(
-                        spotlight,
-                        ritual
-                );
-    }
-
-    private BookEntryModel makeCraftMaridMinerEntry(CategoryEntryMap entryMap, char icon) {
-        this.context().entry("craft_marid_miner");
-
-        this.context().page("spotlight");
-        var spotlight = BookSpotlightPageModel.create()
-                .withItem(Ingredient.of(OccultismItems.MINER_MARID_MASTER.get()))
-                .withText(this.context().pageText());
-
-        this.context().page("ritual");
-        var ritual = BookRitualRecipePageModel.create()
-                .withRecipeId1(this.modLoc("ritual/craft_miner_marid_master"));
-
-        return BookEntryModel.create(this.modLoc(this.context().categoryId() + "/" + this.context().entryId()), this.context().entryName())
-                .withIcon(OccultismItems.MINER_MARID_MASTER.get())
-                .withLocation(entryMap.get(icon))
-                .withPages(
-                        spotlight,
-                        ritual
-                );
-    }
-
-    private BookEntryModel makeCraftSatchelEntry(CategoryEntryMap entryMap, char icon) {
-        this.context().entry("craft_satchel");
-
-        this.context().page("spotlight");
-        var spotlight = BookSpotlightPageModel.create()
-                .withItem(Ingredient.of(OccultismItems.SATCHEL.get()))
-                .withText(this.context().pageText());
-
-        this.context().page("ritual");
-        var ritual = BookRitualRecipePageModel.create()
-                .withRecipeId1(this.modLoc("ritual/craft_satchel"));
-
-        return BookEntryModel.create(this.modLoc(this.context().categoryId() + "/" + this.context().entryId()), this.context().entryName())
-                .withIcon(OccultismItems.SATCHEL.get())
-                .withLocation(entryMap.get(icon))
-                .withPages(
-                        spotlight,
-                        ritual
-                );
-    }
-
-    private BookEntryModel makeCraftSoulGemEntry(CategoryEntryMap entryMap, char icon) {
-        this.context().entry("craft_soul_gem");
-
-        this.context().page("spotlight");
-        var spotlight = BookSpotlightPageModel.create()
-                .withItem(Ingredient.of(OccultismItems.SOUL_GEM_ITEM.get()))
-                .withText(this.context().pageText());
-
-        this.context().page("usage");
-        var usage = BookTextPageModel.create()
-                .withTitle(this.context().pageTitle())
-                .withText(this.context().pageText());
-
-        this.context().page("ritual");
-        var ritual = BookRitualRecipePageModel.create()
-                .withRecipeId1(this.modLoc("ritual/craft_soul_gem"));
-
-        return BookEntryModel.create(this.modLoc(this.context().categoryId() + "/" + this.context().entryId()), this.context().entryName())
-                .withIcon(OccultismItems.SOUL_GEM_ITEM.get())
-                .withLocation(entryMap.get(icon))
-                .withPages(
-                        spotlight,
-                        usage,
-                        ritual
-                );
-    }
-
-    private BookEntryModel makeCraftFamiliarRingEntry(CategoryEntryMap entryMap, char icon) {
-        this.context().entry("craft_familiar_ring");
-
-        this.context().page("spotlight");
-        var spotlight = BookSpotlightPageModel.create()
-                .withItem(Ingredient.of(OccultismItems.FAMILIAR_RING.get()))
-                .withText(this.context().pageText());
-
-        this.context().page("usage");
-        var usage = BookTextPageModel.create()
-                .withTitle(this.context().pageTitle())
-                .withText(this.context().pageText());
-
-        this.context().page("ritual");
-        var ritual = BookRitualRecipePageModel.create()
-                .withRecipeId1(this.modLoc("ritual/craft_familiar_ring"));
-
-        return BookEntryModel.create(this.modLoc(this.context().categoryId() + "/" + this.context().entryId()), this.context().entryName())
-                .withIcon(OccultismItems.FAMILIAR_RING.get())
-                .withLocation(entryMap.get(icon))
-                .withPages(
-                        spotlight,
-                        usage,
-                        ritual
-                );
-    }
-
-    private BookEntryModel makeCraftOtherworldGogglesEntry(CategoryEntryMap entryMap, char icon) {
-        this.context().entry("craft_otherworld_goggles");
-
-        this.context().page("goggles_spotlight");
-        var gogglesSpotlight = BookSpotlightPageModel.create()
-                .withItem(Ingredient.of(OccultismItems.OTHERWORLD_GOGGLES.get()))
-                .withText(this.context().pageText());
-
-        this.context().page("goggles_more");
-        var gogglesMore = BookTextPageModel.create()
-                .withText(this.context().pageText());
-
-        this.context().page("lenses_spotlight");
-        var lensesSpotlight = BookSpotlightPageModel.create()
-                .withItem(Ingredient.of(OccultismItems.LENSES.get()))
-                .withText(this.context().pageText());
-
-        this.context().page("lenses_more");
-        var lensesMore = BookTextPageModel.create()
-                .withTitle(this.context().pageTitle())
-                .withText(this.context().pageText());
-
-        this.context().page("lenses_recipe");
-        var lensesRecipe = BookCraftingRecipePageModel.create()
-                .withRecipeId1(this.modLoc("crafting/lenses"));
-
-        this.context().page("ritual");
-        var ritual = BookRitualRecipePageModel.create()
-                .withRecipeId1(this.modLoc("ritual/craft_infused_lenses"));
-
-        this.context().page("goggles_recipe");
-        var gogglesRecipe = BookCraftingRecipePageModel.create()
-                .withRecipeId1(this.modLoc("crafting/goggles"));
-
-        return BookEntryModel.create(this.modLoc(this.context().categoryId() + "/" + this.context().entryId()), this.context().entryName())
-                .withIcon(OccultismItems.OTHERWORLD_GOGGLES.get())
-                .withLocation(entryMap.get(icon))
-                .withPages(
-                        gogglesSpotlight,
-                        gogglesMore,
-                        lensesSpotlight,
-                        lensesMore,
-                        lensesRecipe,
-                        ritual,
-                        gogglesRecipe
-                );
-    }
-
-    private BookEntryModel makeCraftWildTrimEntry(CategoryEntryMap entryMap, char icon) {
-        this.context().entry("craft_wild_trim");
-
-        this.context().page("spotlight");
-        var spotlight = BookSpotlightPageModel.create()
-                .withItem(Ingredient.of(Items.WILD_ARMOR_TRIM_SMITHING_TEMPLATE))
-                .withText(this.context().pageText());
-
-        this.context().page("ritual");
-        var ritual = BookRitualRecipePageModel.create()
-                .withRecipeId1(this.modLoc("ritual/craft_wild_trim"));
-
-        return BookEntryModel.create(this.modLoc(this.context().categoryId() + "/" + this.context().entryId()), this.context().entryName())
-                .withIcon(Items.WILD_ARMOR_TRIM_SMITHING_TEMPLATE)
-                .withLocation(entryMap.get(icon))
-                .withPages(
-                        spotlight,
-                        ritual
-                );
-    }
-
-    //endregion
-
     //region Possession Rituals
     private BookCategoryModel makePossessionRitualsSubcategory() {
         this.context().category("possession_rituals");
@@ -5460,23 +4840,25 @@ public class OccultismBookProvider extends SingleBookSubProvider {
         storageStabilizer.withParent(BookEntryParentModel.create(storageController.getId()));
 
         this.context().category("crafting_rituals"); //re-use existing entries
-        var craftStabilizerTier1 = this.makeCraftStabilizerTier1Entry(entryMap, '1');
+        var bindingRitualsCategory = new BindingRitualsCategory(this);
+
+        var craftStabilizerTier1 = bindingRitualsCategory.makeCraftStabilizerTier1Entry(entryMap, '1');
         craftStabilizerTier1.withParent(BookEntryParentModel.create(storageStabilizer.getId()));
-        var craftStabilizerTier2 = this.makeCraftStabilizerTier2Entry(entryMap, '2');
+        var craftStabilizerTier2 = bindingRitualsCategory.makeCraftStabilizerTier2Entry(entryMap, '2');
         craftStabilizerTier2.withParent(BookEntryParentModel.create(
                 ResourceLocation.fromNamespaceAndPath(
                         craftStabilizerTier1.getId().getNamespace(),
                         "storage/" + craftStabilizerTier1.getId().getPath()
                 )
         ));
-        var craftStabilizerTier3 = this.makeCraftStabilizerTier3Entry(entryMap, '3');
+        var craftStabilizerTier3 = bindingRitualsCategory.makeCraftStabilizerTier3Entry(entryMap, '3');
         craftStabilizerTier3.withParent(BookEntryParentModel.create(
                 ResourceLocation.fromNamespaceAndPath(
                         craftStabilizerTier2.getId().getNamespace(),
                         "storage/" + craftStabilizerTier2.getId().getPath()
                 )
         ));
-        var craftStabilizerTier4 = this.makeCraftStabilizerTier4Entry(entryMap, '4');
+        var craftStabilizerTier4 = bindingRitualsCategory.makeCraftStabilizerTier4Entry(entryMap, '4');
         craftStabilizerTier4.withParent(BookEntryParentModel.create(
                 ResourceLocation.fromNamespaceAndPath(
                         craftStabilizerTier3.getId().getNamespace(),
@@ -5484,9 +4866,9 @@ public class OccultismBookProvider extends SingleBookSubProvider {
                 )
         ));
 
-        var craftStableWormhole = this.makeCraftStableWormholeEntry(entryMap, 'w');
+        var craftStableWormhole = bindingRitualsCategory.makeCraftStableWormholeEntry(entryMap, 'w');
         craftStableWormhole.withParent(BookEntryParentModel.create(storageController.getId()));
-        var craftStorageRemote = this.makeCraftStorageRemoteEntry(entryMap, 'r');
+        var craftStorageRemote = bindingRitualsCategory.makeCraftStorageRemoteEntry(entryMap, 'r');
         craftStorageRemote.withParent(BookEntryParentModel.create(
                 ResourceLocation.fromNamespaceAndPath(
                         craftStableWormhole.getId().getNamespace(),
